requirements.txt
Flask==1.1.1
fastai==1.0.61
opencv-python==4.2.0.34
numpy==1.18.1
scikit-image==0.16.2
Pillow==7.0.0
Werkzeug==1.0.1
gunicorn==20.0.4
